% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGeno.R
\name{filterGeno}
\alias{filterGeno}
\title{filterGeno function}
\usage{
filterGeno(M, MARGIN = 2, MAF = 0.01, MISS = 0.5, HET = 0.05,
  returnMatrix = TRUE, returnStats = FALSE, returnImputed = FALSE,
  rmDupl = FALSE, rmDuplBy = NULL, maxGdiff = 0, maxiter = 10,
  checkAllMiss = FALSE, returnBool = FALSE)
}
\arguments{
\item{M}{[value]}

\item{MARGIN}{[value]. Default is 2}

\item{MAF}{[value]. Default is 0.01}

\item{MISS}{[value]. Default is 0.5}

\item{HET}{[value]. Default is 0.05}

\item{returnMatrix}{[value]. Default is TRUE}

\item{returnStats}{[value]. Default is FALSE}

\item{returnImputed}{[value]. Default is FALSE}

\item{rmDupl}{[value]. Default is FALSE}

\item{rmDuplBy}{[value]. Default is NULL}

\item{maxGdiff}{[value]. Default is 0}

\item{maxiter}{[value]. Default is 10}

\item{checkAllMiss}{[value]. Default is FALSE}

\item{returnBool}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
