% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupacToNumeric.R
\name{iupacToNumeric}
\alias{iupacToNumeric}
\title{iupacToNumeric function}
\usage{
iupacToNumeric(X, NAsymbol = "N", snpMargin = 2, checkX = TRUE,
  returnMAF = FALSE, includePM = TRUE)
}
\arguments{
\item{X}{data.frame or matrix of IUPAC genotype calls}

\item{NAsymbol}{character for missing call symbol. Default is "N"}

\item{snpMargin}{integer. margin for snps. the other margin should be for individuals. 1 for rows, 2 for columns. Default is 2}

\item{checkX}{logical. Should X be checked for > 2 alleles per site? Default is TRUE}

\item{returnMAF}{logical. Should tables of allele and genotype frequencies be returned? Default is FALSE}

\item{includePM}{logical. Should '+' and '-' symbols be processed? Default is TRUE.}
}
\value{
data.frame or matrix (depending on input) of genotype scores as numeric values.
}
\description{
function to (do something)
}
\details{
This funciton expects single character IUPAC scores for each genotype call. FOr example, an 'AA' call would be 'A', and an 'AT' call would be 'W'
}
\examples{
none
}
