% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCplot.R
\name{PCplot}
\alias{PCplot}
\title{PCplot function}
\usage{
PCplot(X, PC = NULL, xflip = FALSE, yflip = FALSE,
  pdfFileName = NULL, title = NULL, height = 7, width = 7,
  widthRatio = TRUE, ...)
}
\arguments{
\item{X}{[value]}

\item{PC}{[value]. Default is NULL}

\item{xflip}{[value]. Default is FALSE}

\item{yflip}{[value]. Default is FALSE}

\item{pdfFileName}{[value]. Default is NULL}

\item{title}{[value]. Default is NULL}

\item{height}{[value]. Default is 7}

\item{width}{[value]. Default is 7}

\item{widthRatio}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
