% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameAlleles.R
\name{sameAlleles}
\alias{sameAlleles}
\title{sameAlleles function}
\usage{
sameAlleles(x, y, returnTables = FALSE)
}
\arguments{
\item{x}{[value]}

\item{y}{[value]}

\item{returnTables}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
