% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toIUPAC.R
\name{toIUPAC}
\alias{toIUPAC}
\title{toIUPAC function}
\usage{
toIUPAC(X, NAsymbol = "---", checkX = TRUE, includePM = TRUE)
}
\arguments{
\item{X}{[value]}

\item{NAsymbol}{[value]. Default is "---"}

\item{checkX}{[value]. Default is TRUE}

\item{includePM}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
