% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toNumeric.R
\name{toNumeric}
\alias{toNumeric}
\title{toNumeric function}
\usage{
toNumeric(x, check = FALSE, hetAltAlleleToHomoMinor = TRUE,
  returnAf = TRUE, includePM = TRUE)
}
\arguments{
\item{x}{[value]}

\item{check}{[value]. Default is FALSE}

\item{hetAltAlleleToHomoMinor}{[value]. Default is TRUE}

\item{returnAf}{[value]. Default is TRUE}

\item{includePM}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
