% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueColWithNA.R
\name{uniqueColWithNA}
\alias{uniqueColWithNA}
\title{uniqueColWithNA function}
\usage{
uniqueColWithNA(X, corThreshold = 1)
}
\arguments{
\item{X}{[value]}

\item{corThreshold}{[value]. Default is 1}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
