% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanRaden.R
\name{vanRaden}
\alias{vanRaden}
\title{vanRaden function}
\usage{
vanRaden(M, coding = NULL, MARGIN = 2, center = TRUE,
  directCalcEpi = FALSE, addDiag = 0.01, returnVar = FALSE,
  returnMAF = FALSE, returnCentered = FALSE, returnImputed = FALSE,
  checkPhase = FALSE, imp = "mean")
}
\arguments{
\item{M}{[value]}

\item{coding}{[value]. Default is NULL}

\item{MARGIN}{[value]. Default is 2}

\item{center}{[value]. Default is TRUE}

\item{directCalcEpi}{[value]. Default is FALSE}

\item{addDiag}{[value]. Default is 0.01}

\item{returnVar}{[value]. Default is FALSE}

\item{returnMAF}{[value]. Default is FALSE}

\item{returnCentered}{[value]. Default is FALSE}

\item{returnImputed}{[value]. Default is FALSE}

\item{checkPhase}{[value]. Default is FALSE}

\item{imp}{[value]. Default is "mean"}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
