% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupacToNumeric.R
\name{iupacToNumeric}
\alias{iupacToNumeric}
\title{iupacToNumeric function}
\usage{
iupacToNumeric(X, NAsymbol = "N", snpMargin = 2, checkX = TRUE,
  returnMAF = FALSE, includePM = TRUE)
}
\arguments{
\item{X}{[value]}

\item{NAsymbol}{[value]. Default is "N"}

\item{snpMargin}{[value]. Default is 2}

\item{checkX}{[value]. Default is TRUE}

\item{returnMAF}{[value]. Default is FALSE}

\item{includePM}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
